third_party/WebKit/LayoutTests/platform/win/fast/dom/Window/new-window-opener-expected.txt

This tests that the different arguments given to a new window are honored.
Take note that menubar is special and always on in Mac but not in Win.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Scrollbar
PASS newWin.scrollbars.visible is true
PASS successfullyParsed is true

TEST COMPLETE

PASS newWin.scrollbars.visible is false
PASS newWin.scrollbars.visible is true
PASS newWin.scrollbars.visible is false


Toolbar
PASS newWin.toolbar.visible is true
PASS newWin.toolbar.visible is false
PASS newWin.toolbar.visible is true
PASS newWin.toolbar.visible is false


Statusbar
PASS newWin.statusbar.visible is true
PASS newWin.statusbar.visible is false
PASS newWin.statusbar.visible is true
PASS newWin.statusbar.visible is false


Menubar
PASS newWin.menubar.visible is true
PASS newWin.menubar.visible is isMenubarAlwaysExpected
PASS newWin.menubar.visible is true
PASS newWin.menubar.visible is isMenubarAlwaysExpected


Locationbar
PASS newWin.locationbar.visible is true
PASS newWin.locationbar.visible is false
PASS newWin.locationbar.visible is true
PASS newWin.locationbar.visible is false
PASS successfullyParsed is true

TEST COMPLETE
